[LF-35/batch-modify] 배치 전체 로직 이동 완료
- DB 분리해야함
- 정상 작동 하는지 확인해야 함
