+++ LiveFeedBatch/build.gradle
@@ -10,6 +10,10 @@ dependencies {
     implementation 'org.seleniumhq.selenium:selenium-java:4.11.0'
     implementation 'io.github.bonigarcia:webdrivermanager:5.4.1'
     implementation 'mysql:mysql-connector-java:8.0.33'
+    implementation 'org.springframework.boot:spring-boot-starter-data-redis'
+    implementation 'org.springframework.boot:spring-boot-starter-data-elasticsearch'
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+
     testImplementation 'org.springframework.batch:spring-batch-test'
     testImplementation 'com.h2database:h2'
     testImplementation 'io.github.bonigarcia:webdrivermanager:5.4.1'
