+++ LiveFeedBatch/src/main/java/com/livefeed/livefeedbatch/batch/common/dto/ItemDto.java
@@ -1,0 +1,9 @@
+package com.livefeed.livefeedbatch.batch.common.dto;
+
+import com.livefeed.livefeedbatch.batch.common.dto.keydto.UrlInfo;
+
+public record ItemDto(
+        UrlInfo urlInfo,
+        ItemDtoInterface itemDtoInterface
+) {
+}
