+++ LiveFeedBatch/src/main/java/com/livefeed/livefeedbatch/batch/common/dto/ItemDtoInterface.java
@@ -1,0 +1,6 @@
+package com.livefeed.livefeedbatch.batch.common.dto;
+
+public interface ItemDtoInterface<T> {
+
+    T getValue();
+}
