+++ LiveFeedBatch/src/main/java/com/livefeed/livefeedbatch/batch/common/dto/keydto/Platform.java
@@ -1,0 +1,5 @@
+package com.livefeed.livefeedbatch.batch.common.dto.keydto;
+
+public enum Platform {
+    NAVER, GOOGLE
+}
