+++ LiveFeedBatch/src/main/java/com/livefeed/livefeedbatch/batch/common/dto/wrtiervaluedto/WriterValue.java
@@ -1,0 +1,13 @@
+package com.livefeed.livefeedbatch.batch.common.dto.wrtiervaluedto;
+
+import com.livefeed.livefeedbatch.batch.common.dto.ItemDtoInterface;
+
+public record WriterValue(
+        String url
+) implements ItemDtoInterface<WriterValue> {
+
+    @Override
+    public WriterValue getValue() {
+        return this;
+    }
+}
