+++ LiveFeedBatch/src/main/java/com/livefeed/livefeedbatch/batch/processor/parser/exception/NoMatchingParserException.java
@@ -1,0 +1,7 @@
+package com.livefeed.livefeedbatch.batch.processor.parser.exception;
+
+public class NoMatchingParserException extends RuntimeException {
+    public NoMatchingParserException(String message) {
+        super(message);
+    }
+}
