+++ LiveFeedBatch/src/main/java/com/livefeed/livefeedbatch/batch/processor/parser/exception/ParsingException.java
@@ -1,0 +1,8 @@
+package com.livefeed.livefeedbatch.batch.processor.parser.exception;
+
+public class ParsingException extends RuntimeException {
+
+    public ParsingException(String message) {
+        super(message);
+    }
+}
