+++ LiveFeedBatch/src/main/java/com/livefeed/livefeedbatch/batch/writer/elasticsearch/configuration/ElasticsearchClientConfig.java
@@ -1,0 +1,16 @@
+package com.livefeed.livefeedbatch.batch.writer.elasticsearch.configuration;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.elasticsearch.client.ClientConfiguration;
+import org.springframework.data.elasticsearch.client.elc.ElasticsearchConfiguration;
+
+@Configuration
+public class ElasticsearchClientConfig extends ElasticsearchConfiguration {
+
+    @Override
+    public ClientConfiguration clientConfiguration() {
+        return ClientConfiguration.builder()
+                .connectedTo("localhost:9200")
+                .build();
+    }
+}
