+++ LiveFeedBatch/src/main/java/com/livefeed/livefeedbatch/batch/writer/elasticsearch/repository/ArticleElasticsearchRepository.java
@@ -1,0 +1,7 @@
+package com.livefeed.livefeedbatch.batch.writer.elasticsearch.repository;
+
+import com.livefeed.livefeedbatch.batch.writer.elasticsearch.entity.ElasticsearchArticle;
+import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
+
+public interface ArticleElasticsearchRepository extends ElasticsearchRepository<ElasticsearchArticle, Long> {
+}
