+++ LiveFeedBatch/src/main/java/com/livefeed/livefeedbatch/batch/writer/rdb/entity/PressCompany.java
@@ -1,0 +1,52 @@
+package com.livefeed.livefeedbatch.batch.writer.rdb.entity;
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.hibernate.annotations.DynamicInsert;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+
+import java.time.LocalDateTime;
+
+@Getter
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@DynamicInsert
+public class PressCompany {
+
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "company_id")
+    private Long id;
+
+    private String companyName;
+
+    @CreatedDate
+    @Column(columnDefinition = "TIMESTAMP DEFAULT CURRENT_TIMESTAMP")
+    private LocalDateTime createdAt;
+
+    @LastModifiedDate
+    @Column(columnDefinition = "TIMESTAMP DEFAULT CURRENT_TIMESTAMP")
+    private LocalDateTime updatedAt;
+
+    @Column(columnDefinition = "boolean default false")
+    private boolean isDeleted;
+
+    @PrePersist
+    public void setDeleteFalse() {
+        this.isDeleted = Boolean.FALSE;
+    }
+
+    @Builder
+    private PressCompany(String companyName) {
+        this.companyName = companyName;
+    }
+
+    public static PressCompany of(String companyName) {
+        return PressCompany.builder()
+                .companyName(companyName)
+                .build();
+    }
+}
