+++ LiveFeedBatch/src/main/java/com/livefeed/livefeedbatch/batch/writer/rdb/repository/CategoryRepository.java
@@ -1,0 +1,15 @@
+package com.livefeed.livefeedbatch.batch.writer.rdb.repository;
+
+
+import com.livefeed.livefeedbatch.batch.common.dto.keydto.Platform;
+import com.livefeed.livefeedbatch.batch.common.dto.keydto.Service;
+import com.livefeed.livefeedbatch.batch.common.dto.keydto.Theme;
+import com.livefeed.livefeedbatch.batch.writer.rdb.entity.Category;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+
+    Optional<Category> findByServiceAndPlatformAndTheme(Service service, Platform platform, Theme theme);
+}
