--- LiveFeedBatch/src/main/java/com/livefeed/livefeedbatch/contentcrawler/parser/dto/BodyDto.java
@@ -1,12 +1,0 @@
-package com.livefeed.livefeedbatch.contentcrawler.parser.dto;
-
-public record BodyDto(
-        String html,
-        String journalistName,
-        String journalistEmail
-) {
-
-    public static BodyDto of(String html, String journalistName, String journalistEmail) {
-        return new BodyDto(html, journalistName, journalistEmail);
-    }
-}
