--- LiveFeedBatch/src/main/java/com/livefeed/livefeedbatch/contentcrawler/parser/dto/HeaderDto.java
@@ -1,15 +1,0 @@
-package com.livefeed.livefeedbatch.contentcrawler.parser.dto;
-
-public record HeaderDto(
-        String html,
-        String articleTitle,
-        String pressCompanyName,
-        String publicationTime,
-        String originArticleUrl
-) {
-
-    public static HeaderDto of(String html, String articleTitle, String pressCompanyName,
-                               String publicationTime, String originArticleUrl) {
-        return new HeaderDto(html, articleTitle, pressCompanyName, publicationTime, originArticleUrl);
-    }
-}
