--- LiveFeedBatch/src/main/java/com/livefeed/livefeedbatch/contentcrawler/parser/exception/NoMatchingParserException.java
@@ -1,7 +1,0 @@
-package com.livefeed.livefeedbatch.contentcrawler.parser.exception;
-
-public class NoMatchingParserException extends RuntimeException {
-    public NoMatchingParserException(String message) {
-        super(message);
-    }
-}
