--- LiveFeedBatch/src/main/java/com/livefeed/livefeedbatch/contentcrawler/parser/exception/ParsingException.java
@@ -1,8 +1,0 @@
-package com.livefeed.livefeedbatch.contentcrawler.parser.exception;
-
-public class ParsingException extends RuntimeException {
-
-    public ParsingException(String message) {
-        super(message);
-    }
-}
