--- LiveFeedBatch/src/main/java/com/livefeed/livefeedbatch/urlcrawler/dto/Platform.java
@@ -1,5 +1,0 @@
-package com.livefeed.livefeedbatch.urlcrawler.dto;
-
-public enum Platform {
-    NAVER, GOOGLE
-}
