--- LiveFeedBatch/src/main/java/com/livefeed/livefeedbatch/urlcrawler/dto/Service.java
@@ -1,5 +1,0 @@
-package com.livefeed.livefeedbatch.urlcrawler.dto;
-
-public enum Service {
-    ARTICLE, BLOG
-}
