--- LiveFeedBatch/src/main/java/com/livefeed/livefeedbatch/urlcrawler/dto/Theme.java
@@ -1,5 +1,0 @@
-package com.livefeed.livefeedbatch.urlcrawler.dto;
-
-public enum Theme {
-    SPORTS, IT
-}
