--- LiveFeedBatch/src/main/java/com/livefeed/livefeedbatch/urlcrawler/dto/UrlInfo.java
@@ -1,8 +1,0 @@
-package com.livefeed.livefeedbatch.urlcrawler.dto;
-
-public record UrlInfo(
-        Service service,
-        Platform platform,
-        Theme theme
-) {
-}
